refactor: move STARK/Atlantic/swiftness code out of saya-core into bin/persistent
saya-core is now pure infrastructure (block ingestor, storage, DA traits,
service primitives). All STARK-specific code (Atlantic HTTP client, mock
prover, sovereign/persistent orchestrators, Piltover STARK settlement, utils)
moves to bin/persistent where it belongs.

- Add optional `snos` feature to saya-core for SovereignPacket and
  impl DataAvailabilityPayload for SnosProof<StarkProof> (orphan rule)
- bin/persistent adds swiftness/integrity/reqwest/piltover/zip as direct deps
- bin/persistent-tee no longer pulls in swiftness via saya-core, fixing the
  size-of v0.1.5 compile error on stable 1.91
