+++ bin/persistent/Cargo.toml
@@ -20,9 +20,29 @@ repository.workspace = true
 
 [dependencies]
 anyhow = { version = "1.0.95", default-features = false }
+base64 = { version = "0.22.1" }
+bigdecimal = { version = "0.3.1", default-features = false }
 cairo-vm = "=2.5.0"
+ciborium = { version = "0.2.2", default-features = false }
 futures-util = { version = "0.3.31", default-features = false }
 generate-pie = { git = "https://github.com/keep-starknet-strange/snos.git", rev = "6cf7040" }
+hex = { version = "0.4.3", default-features = false }
+integrity = { git = "https://github.com/chudkowsky/integrity-rs.git", rev = "9729be1", default-features = false, features = ["recursive_with_poseidon", "keccak_160_lsb", "stone6"] }
+num-traits = { version = "0.2.19", default-features = false }
+piltover = { package = "piltover", git = "https://github.com/cartridge-gg/piltover.git", rev = "67e65b8928b7ee3c2c188bf36c6b9eddc14addb2" }
+cainome = { version = "0.10.1", features = ["abigen-rs"] }
+reqwest = { version = "0.12.12", default-features = false, features = ["json", "multipart", "rustls-tls"] }
+serde = { version = "1.0.217", default-features = false, features = ["derive"] }
+serde_json = { version = "1.0.134", default-features = false }
+starknet-crypto = "0.8.1"
+swiftness = { git = "https://github.com/chudkowsky/swiftness", rev = "e07d185", default-features = false, features = ["recursive_with_poseidon", "keccak_160_lsb", "stone6"] }
+swiftness_air = { git = "https://github.com/chudkowsky/swiftness", rev = "e07d185", default-features = false, features = ["recursive_with_poseidon", "keccak_160_lsb", "stone6"] }
+swiftness_commitment = { git = "https://github.com/chudkowsky/swiftness", rev = "e07d185", default-features = false }
+swiftness_fri = { git = "https://github.com/chudkowsky/swiftness", rev = "e07d185", default-features = false }
+swiftness_pow = { git = "https://github.com/chudkowsky/swiftness", rev = "e07d185", default-features = false }
+swiftness_stark = { git = "https://github.com/chudkowsky/swiftness", rev = "e07d185", default-features = false, features = ["recursive_with_poseidon", "keccak_160_lsb", "stone6"] }
+thiserror = "2.0.12"
+zip = { version = "2.2.2", default-features = false, features = ["deflate"] }
 clap = { version = "4.5.23", default-features = false, features = [
     "derive",
     "env",
@@ -30,7 +50,7 @@ clap = { version = "4.5.23", default-features = false, features = [
 ] }
 env_logger = { version = "0.11.6", features = ["unstable-kv"] }
 log = { version = "0.4.22", features = ["kv"] }
-saya-core = { path = "../../saya/core" }
+saya-core = { path = "../../saya/core", features = ["snos"] }
 starknet = "0.17.0"
 starknet_api = { git = "https://github.com/karnotxyz/sequencer", rev = "e04617e0581d5ec93a035ec0e15419b4c201b629" }
 starknet-types-core = { version = "0.2.1", default-features = false }
