+++ bin/persistent/src/main.rs
@@ -6,6 +6,13 @@
 use anyhow::Result;
 use clap::{Parser, Subcommand};
 
+mod error;
+mod utils;
+mod atlantic;
+mod mock;
+mod settlement;
+mod orchestrator;
+
 mod sovereign;
 use sovereign::Sovereign;
