+++ bin/persistent/src/mock/mod.rs
