+++ bin/persistent/src/orchestrator/sovereign.rs
@@ -3,7 +3,7 @@ use log::{debug, info};
 use swiftness_stark::types::StarkProof;
 use tokio::sync::mpsc::Receiver;
 
-use crate::{
+use saya_core::{
     block_ingestor::{BlockInfo, BlockIngestor, BlockIngestorBuilder},
     data_availability::{
         DataAvailabilityBackend, DataAvailabilityBackendBuilder, DataAvailabilityCursor,
