+++ saya/core/src/data_availability/mod.rs
@@ -2,6 +2,7 @@ use anyhow::Result;
 use serde::{Deserialize, Serialize};
 use starknet::core::types::Felt;
 use starknet::core::types::StateUpdate;
+#[cfg(feature = "snos")]
 use swiftness_stark::types::StarkProof;
 use tokio::sync::mpsc::{Receiver, Sender};
 
@@ -11,7 +12,9 @@ pub use celestia::{CelestiaDataAvailabilityBackend, CelestiaDataAvailabilityBack
 mod noop;
 pub use noop::{NoopDataAvailabilityBackend, NoopDataAvailabilityBackendBuilder};
 
-use crate::{prover::SnosProof, service::Daemon};
+#[cfg(feature = "snos")]
+use crate::prover::SnosProof;
+use crate::service::Daemon;
 
 pub trait DataAvailabilityBackendBuilder {
     type Backend: DataAvailabilityBackend;
@@ -51,6 +54,7 @@ pub struct DataAvailabilityPacketContext {
 
 /// Data made available in `sovereign` mode, which contains the full `snos` proof to be verified
 /// off-chain.
+#[cfg(feature = "snos")]
 #[derive(Debug, Serialize, Deserialize)]
 pub struct SovereignPacket {
     /// Pointer to the previous [`SovereignPacket`].
@@ -60,21 +64,6 @@ pub struct SovereignPacket {
 }
 
 /// Data made available in `persistent` mode, containing only the `snos` output.
-///
-/// No STARK proof needs to be made available as proofs are supposedly verified in a decentralized
-/// manner in `persistent` mode.
-///
-/// Note that depending on the settlement layer, the exact data to be made available could be
-/// different. However, since currently only one settlement implementation (i.e. `piltover`) is
-/// available, this type is hard-coded to tailor for that specific implementation.
-///
-/// Also note that, technically speaking, no data needs to be made available for the current
-/// `piltover` implementation, as its `update_state` entrypoint takes full `snos` output, which
-/// contains the full state diff needed to reconstruct network state. However, this `piltover`
-/// behaviour is considered suboptimal and will eventually be changed to take only the digest of the
-/// `snos` output, at which point something (not necessarily the full `snos` output; probably just
-/// the state diff section) needs to be made available anyway, so we might as well just keep the DA
-/// posting mechanism in place for now.
 #[derive(Debug, Serialize, Deserialize)]
 pub struct PersistentPacket {
     pub state_update: Option<StateUpdate>,
