+++ saya/core/src/orchestrator/mod.rs
@@ -1,9 +1,3 @@
-mod sovereign;
-pub use sovereign::{SovereignOrchestrator, SovereignOrchestratorBuilder};
-
-mod persistent;
-pub use persistent::{PersistentOrchestrator, PersistentOrchestratorBuilder};
-
 mod persistent_tee;
 pub use persistent_tee::{PersistentTeeOrchestrator, PersistentTeeOrchestratorBuilder};
