+++ saya/core/src/storage/sql_lite/storage.rs
@@ -338,7 +338,7 @@ impl PersistantStorage for SqliteDb {
 
 #[cfg(test)]
 mod tests {
-    use integrity::Felt;
+    use starknet_types_core::Felt;
     use starknet::core::types::StateDiff;
 
     use crate::storage::sql_lite::IN_MEMORY_DB;
